Added Game Loop
Added a basic game loop. The play button in the main menu can now load the level scene. The players score is tracked. Currently the player gets points when delivering a package. A timer has been added which counts down. An end scene is loaded when the timer reaches zero.